Init commit
Homework CSD322

get - The get command is how ftp-client retrieves a file from ftp-server. The filename argument to the command is required. If the get command does not include a file name, or the named file does not exist, ftp-server responds with the string, "...huh?...". Otherwise, ftp-server opens the named file and sends it over the socket to ftp-client.

put - This command is used by ftp-client to send a file to ftp-server. The filename argument is required. If the put command does not include a file name, ftp-server responds with the string, "...huh?...". Otherwise, ftp-server creates a local file with the name specified in the put command, and stores the incoming data in it.

dir - The ftp-server responds to this command by getting a directory listing, and sending it over the socket to ftp-client.

help - ftp-server responds to this command by send a list of the commands with a brief explanation of each.

bye - ftp-server responds by closing the server socket, and then waiting for another connection.

del - the ftp-server responds to this command by deleting the named file from it's current directory. If the file does not exist, ftp-server responds with the string "...huh?..."

rename- the ftp-server responds to this command by renaming the named file in its current directory to the new file name. If the file does not exist, or if the second file name is not provided, ftp-server responds with the string "...huh?..."